Revert

+++ pkg/api/client/service_group.go
@@ -13,7 +13,7 @@ func NewServiceGroupClient(c ScoretrakClient) service_group.Serv {
 	return &ServiceGroupClient{c}
 }
 
-func (s ServiceGroupClient) Delete(id uint32) error {
+func (s ServiceGroupClient) Delete(id uint64) error {
 	return s.s.GenericDelete(fmt.Sprintf("/service_group/%d", id))
 }
 
@@ -26,7 +26,7 @@ func (s ServiceGroupClient) GetAll() ([]*service_group.ServiceGroup, error) {
 	return sg, nil
 }
 
-func (s ServiceGroupClient) GetByID(id uint32) (*service_group.ServiceGroup, error) {
+func (s ServiceGroupClient) GetByID(id uint64) (*service_group.ServiceGroup, error) {
 	sg := &service_group.ServiceGroup{}
 	err := s.s.GenericGet(sg, fmt.Sprintf("/service_group/%d", id))
 	if err != nil {
@@ -43,7 +43,7 @@ func (s ServiceGroupClient) Update(u *service_group.ServiceGroup) error {
 	return s.s.GenericUpdate(u, fmt.Sprintf("/service_group/%d", u.ID))
 }
 
-func (s ServiceGroupClient) Redeploy(id uint32) error {
+func (s ServiceGroupClient) Redeploy(id uint64) error {
 	err := s.s.GenericGet(nil, fmt.Sprintf("/service_group/%d/redeploy", id))
 	if err != nil {
 		return err
